requirements.txt
tensorflow==2.16.1
opencv-python>=4.8.0
numpy>=1.26.0
flask>=2.3.3
pillow>=10.0.0
mtcnn>=0.1.1
efficientnet>=1.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0 